+++ NEWS.md
@@ -1,14 +1,18 @@
 # Release notes
 
-## 1.5.0 -- not released yet
+## 1.5.0 -- 2025-06-19
 
 * Updated the addon for Minecraft Bedrock 1.21.81.
 * Fixed custom commands `;qmine prefs` and `;qmine admin` that have stopped
-  working since MCBE 1.21.80.
+  working since MCBE 1.21.80. Reported by TELDAX-2
 * Migrated to native custom commands introduced in MCBE 1.21.80. The
   per-player preferences UI can now be opened via `/qmine:prefs`, and the
   admin UI can be opened via `/qmine:admin`. Legacy commands still work,
-  but they are deprecated and will probably be removed in the future.
+  but they are deprecated and will probably be removed in the future. (The
+  migration was my plan, but also suggested separately by XEnzoX9808)
+* Added [Leaf Litter](https://minecraft.wiki/w/Leaf_Litter) to the category
+  **Plants and Crops**. It requires shears to quick-mine, and does not
+  consume durability. Suggested by captaingreed477
 
 ## 1.4.0 -- 2025-03-27